Script/requirements.txt.txt
csv
opencv-python
numpy
face-recognition
mysql-connector-python
